+++ game_patch/rf/ai.h
@@ -225,6 +225,7 @@ namespace rf
         addr_as_ref<void(AiInfo* ai_info, AiMode mode, int mode_param_0, int mode_param_1)>(0x00407E20);
     static auto& ai_set_submode =
         addr_as_ref<void(AiInfo* ai_info, AiSubmode submode)>(0x00407E80);
+    static auto& ai_go_berserk = addr_as_ref<void(AiInfo* ai_info, float berserk_time_second, bool force)>(0x004089F0);
     static auto& ai_enter_attack_mode = addr_as_ref<void(AiInfo* ai_info)>(0x00408AC0);
     static auto& ai_acquire_new_target = addr_as_ref<void(AiInfo* ai_info)>(0x00403A80);
     static auto& ai_path_locate_pos = addr_as_ref<void(Vector3* pos, float radius, float height, bool walk_only,